Create Briscola score keeper MainActivity for android basic challenge
This code is an exercise for project 2 . " Briscola" is a Sicily card game that as five type of points: woman +2; horse+3; king+4; three+10 and ace+11. The basic version of the game is for two players. The total amount of point is 120, wins who get over 60 points. But there is another version that is for 5 players. Every player makes is guess on the result, a value from 61 to 120. The higher is the capitan of the first team. He choose the second of his team. The remain three are the componets of the second team.